+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [1.0.0] - 2022-05-16
 - Changed: drop symfony 3 support
 - Changed: refactoring
+- Added: focus on video element after confirming protected code
 
 ## [0.12.0] - 2022-05-16
 - Updated: symfony/http-kernel dep up to 5.4